"Set `name` and `matrix` properties for Ins and Gates"
Set 'name' as an abstract property for any abstract base class like Instruction and as a class attribute for concrete class like XGate. Some logic for adding `matrix` to QuantumCircuit and its subclasses. Matrices of gate-cls with paras are properties decorated by @property, which update matrices automatically when parameters updated.